Add files via upload
This is a school management system which aims to create a liaison between teachers and students. It does this by offering a platform through which both parties can efficiently track the data relevant to them. Through a Main Home Page, the user can create an account, following which they are assigned specific user ID depending on if they are a student or a teacher. After logging in with this user ID and password, they are redirected to the relevant portal as illustrated below.
The teacher portal offers the following functions:

• View details of all students and teachers registered on the site
• Enter / Edit student marks for the subject they teach
• View marks of all students registered on the site

The student portal offers the following features:
• View marks of student
• View details of all students registered on the site

This project has also been closely integrated with the MySQL Database Management System, to make an inventory of data in three tables - teacher details, student details and student marks.